requirements/base.txt
Django>=4.2.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.3
django-extensions>=3.2.3
django-environ>=0.11.2
django-modeltranslation>=0.18.11
psycopg2-binary>=2.9.9
celery>=5.3.4
redis>=5.0.1
Pillow>=10.0.0
PyJWT>=2.8.0
djangorestframework-simplejwt>=5.3.0
django-polymorphic>=3.1.0
django-mptt>=0.14.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
whitenoise>=6.6.0
gunicorn>=21.2.0